samples: Bluetooth: remove peripheral_hids
The peripheral_hogp sample added by the previous commit covers what this
sample demonstrates, and does it with the HID Service rather than a copy
of it, so keep one HID Device sample in the tree.

The service this sample declares is also not one to keep as the example a
HID Device is written from. Its HID Information, Report Map, Report
Reference descriptor and HID Control Point are readable or writable on an
unencrypted link, which the HID over GATT Profile does not allow of any
HID Service characteristic. A read of the Report returns zero octets, so
a Host that asks for the current value never gets one. The HID Control
Point stores whatever is written to it, including the reserved values.
The Report is notified through a hardcoded attribute table index, and to
a value the CCC of any Host enables for every Host. The HID Information
declares specification version 0x0000.

Carry over the board overlay, which routes the interrupts the radio of
rv32m1_vega needs and has nothing to do with the sample, and the option
to require an authenticated link. The option used to raise the attribute
permissions of the Report, which the service does not offer because the
profile requires an encrypted link and does not ask for
Man-In-The-Middle protection; it now requests Security Level 3 from the
connection instead, which an application can do without the service
having to expose anything. Its default changes from enabled to disabled,
so that what the profile requires is what the sample does by default.

The documentation of the removed sample redirects to the new one.
