--- samples/bluetooth/peripheral_hids/prj.conf
@@ -1,19 +1,0 @@
-# Increased stack due to settings API usage
-CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=2048
-
-CONFIG_BT=y
-CONFIG_LOG=y
-CONFIG_BT_SMP=y
-CONFIG_BT_PERIPHERAL=y
-CONFIG_BT_DIS=y
-CONFIG_BT_BAS=y
-CONFIG_BT_DEVICE_NAME="Test HoG mouse"
-CONFIG_BT_DEVICE_APPEARANCE=962
-
-CONFIG_BT_SETTINGS=y
-CONFIG_FLASH=y
-CONFIG_FLASH_MAP=y
-CONFIG_NVS=y
-CONFIG_SETTINGS=y
-
-CONFIG_GPIO=y
