--- samples/bluetooth/peripheral_hids/src/hog.h
@@ -1,21 +1,0 @@
-/** @file
- *  @brief HoG Service sample
- */
-
-/*
- * Copyright (c) 2016 Intel Corporation
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-void hog_init(void);
-
-void hog_button_loop(void);
-
-#ifdef __cplusplus
-}
-#endif
