--- samples/bluetooth/peripheral_hids/tests.yaml
@@ -1,18 +1,0 @@
-sample:
-  name: Bluetooth Peripheral HIDs
-  description: Bluetooth Low Energy HID-over-GATT service sample
-tests:
-  sample.bluetooth.peripheral_hids:
-    harness: bluetooth
-    platform_allow: qemu_x86
-    tags: bluetooth
-    integration_platforms:
-      - qemu_x86
-  sample.bluetooth.peripheral_hids.no_authentication:
-    harness: bluetooth
-    extra_configs:
-      - CONFIG_SAMPLE_BT_USE_AUTHENTICATION=n
-    platform_allow: qemu_x86
-    tags: bluetooth
-    integration_platforms:
-      - qemu_x86
