+++ samples/bluetooth/peripheral_hogp/boards/rv32m1_vega_openisa_rv32m1_ri5cy.overlay
