+++ samples/bluetooth/peripheral_hogp/tests.yaml
@@ -10,3 +10,13 @@ tests:
       - nrf52840dk/nrf52840
     integration_platforms:
       - native_sim/native/64
+  sample.bluetooth.peripheral_hogp.authentication:
+    harness: bluetooth
+    tags: bluetooth
+    extra_configs:
+      - CONFIG_SAMPLE_BT_USE_AUTHENTICATION=y
+    platform_allow:
+      - native_sim/native/64
+      - nrf52840dk/nrf52840
+    integration_platforms:
+      - native_sim/native/64
